+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## UNRELEASED
+
+### Adds
+
+* Adds a checkbox to select and deselect all related document types in the export modal.
+
 ## 3.2.1 (2025-05-14)
 
 ### Fixes